adjusting mouse interactions
loaded draft model to the site to begin wireframing. changed the design a bit (changed colors and titles) and working on the map, but the tooltip is currently messed up. i want selection area to be highlighted on mouseover but when i did it for polygons it messed everything else up